+++ game/exceptions.py
@@ -1,5 +1,5 @@
 class InvalidMove(Exception):
-    message = "Movimieto de pieza invalido"
+    message = "Movimiento de pieza invalido"
     def __str__(self):
         return self.message