+++ tests/test_queen.py
@@ -4,22 +4,20 @@ from game.pieces import Queen
 from game.pieces import Pawn
 class TestQueen(unittest.TestCase):
 
-    def setUp(self):
-        self.board = Board()  # Esto lo hice para que las piezas con el tablero inicial 
-        for row in range(8):  # no me molesten al momento de hacer las pruebas de movimientos
-            for col in range(8):
-                self.board.set_piece(row, col, None)
+    
 
     def test_str(self):
-        queen = Queen("WHITE", self.board)
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
         self.assertEqual(
             str(queen),
             "♛",
         )
 
     def test_move_vertical_up(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
         possibles = queen.possible_movement_vertical_up(4, 4)
         self.assertEqual(
             possibles,
@@ -27,9 +25,10 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_vertical_up_with_obstacle(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(2, 4, queen)
-        self.board.set_piece(4, 4, Pawn("WHITE", self.board))
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(2, 4, queen)
+        board.set_piece(4, 4, Pawn("WHITE", board))
         possibles = queen.possible_movement_vertical_up(2, 4)
         self.assertEqual(
             possibles,
@@ -37,9 +36,10 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_vertical_up_attack(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(1, 4, queen)
-        self.board.set_piece(3, 4, Pawn("BLACK", self.board))
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(1, 4, queen)
+        board.set_piece(3, 4, Pawn("BLACK", board))
         possibles = queen.possible_movement_vertical_up(1, 4)
         self.assertEqual(
             possibles,
@@ -47,8 +47,9 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_vertical_down(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
         possibles = queen.possible_movement_vertical_down(4, 4)
         self.assertEqual(
             possibles,
@@ -56,9 +57,10 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_vertical_down_with_obstacle(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(6, 4, queen)
-        self.board.set_piece(2, 4, Pawn("WHITE", self.board))
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(6, 4, queen)
+        board.set_piece(2, 4, Pawn("WHITE", board))
         possibles = queen.possible_movement_vertical_down(6, 4)
         self.assertEqual(
             possibles,
@@ -66,9 +68,10 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_vertical_down_attack(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(6, 4, queen)
-        self.board.set_piece(2, 4, Pawn("BLACK", self.board))
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(6, 4, queen)
+        board.set_piece(2, 4, Pawn("BLACK", board))
         possibles = queen.possible_movement_vertical_down(6, 4)
         self.assertEqual(
             possibles,
@@ -76,8 +79,9 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_horizontal_right(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
         possibles = queen.possible_movement_horizontal_right(4, 4)
         self.assertEqual(
             possibles,
@@ -85,9 +89,10 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_horizontal_right_with_obstacle(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
-        self.board.set_piece(4, 6, Pawn("WHITE", self.board))
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
+        board.set_piece(4, 6, Pawn("WHITE", board))
         possibles = queen.possible_movement_horizontal_right(4, 4)
         self.assertEqual(
             possibles,
@@ -95,9 +100,10 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_horizontal_right_attack(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
-        self.board.set_piece(4, 6, Pawn("BLACK", self.board))
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
+        board.set_piece(4, 6, Pawn("BLACK", board))
         possibles = queen.possible_movement_horizontal_right(4, 4)
         self.assertEqual(
             possibles,
@@ -105,8 +111,9 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_horizontal_left(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
         possibles = queen.possible_movement_horizontal_left(4, 4)
         self.assertEqual(
             possibles,
@@ -114,9 +121,10 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_horizontal_left_with_obstacle(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
-        self.board.set_piece(4, 2, Pawn("WHITE", self.board))
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
+        board.set_piece(4, 2, Pawn("WHITE", board))
         possibles = queen.possible_movement_horizontal_left(4, 4)
         self.assertEqual(
             possibles,
@@ -124,9 +132,10 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_horizontal_left_attack(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
-        self.board.set_piece(4, 2, Pawn("BLACK", self.board))
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
+        board.set_piece(4, 2, Pawn("BLACK", board))
         possibles = queen.possible_movement_horizontal_left(4, 4)
         self.assertEqual(
             possibles,
@@ -134,8 +143,9 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_diagonal_up_left(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
         possibles = queen.possible_positions_diagonal_up_left(4, 4)
         self.assertEqual(
             possibles,
@@ -143,9 +153,10 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_diagonal_up_left_with_obstacle(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
-        self.board.set_piece(6, 2, Pawn("WHITE", self.board))
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
+        board.set_piece(6, 2, Pawn("WHITE", board))
         possibles = queen.possible_positions_diagonal_up_left(4, 4)
         self.assertEqual(
             possibles,
@@ -153,9 +164,10 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_diagonal_up_left_attack(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
-        self.board.set_piece(6, 2, Pawn("BLACK", self.board))
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
+        board.set_piece(6, 2, Pawn("BLACK", board))
         possibles = queen.possible_positions_diagonal_up_left(4, 4)
         self.assertEqual(
             possibles,
@@ -163,8 +175,9 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_diagonal_up_right(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
         possibles = queen.possible_positions_diagonal_up_right(4, 4)
         self.assertEqual(
             possibles,
@@ -172,9 +185,10 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_diagonal_up_right_with_obstacle(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
-        self.board.set_piece(6, 6, Pawn("WHITE", self.board))
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
+        board.set_piece(6, 6, Pawn("WHITE", board))
         possibles = queen.possible_positions_diagonal_up_right(4, 4)
         self.assertEqual(
             possibles,
@@ -182,9 +196,10 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_diagonal_up_right_attack(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
-        self.board.set_piece(6, 6, Pawn("BLACK", self.board))
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
+        board.set_piece(6, 6, Pawn("BLACK", board))
         possibles = queen.possible_positions_diagonal_up_right(4, 4)
         self.assertEqual(
             possibles,
@@ -192,8 +207,9 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_diagonal_down_right(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
         possibles = queen.possible_positions_diagonal_down_right(4, 4)
         self.assertEqual(
             possibles,
@@ -201,9 +217,10 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_diagonal_down_right_with_obstacle(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
-        self.board.set_piece(2, 6, Pawn("WHITE", self.board))
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
+        board.set_piece(2, 6, Pawn("WHITE", board))
         possibles = queen.possible_positions_diagonal_down_right(4, 4)
         self.assertEqual(
             possibles,
@@ -211,9 +228,10 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_diagonal_down_right_attack(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
-        self.board.set_piece(2, 6, Pawn("BLACK", self.board))
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
+        board.set_piece(2, 6, Pawn("BLACK", board))
         possibles = queen.possible_positions_diagonal_down_right(4, 4)
         self.assertEqual(
             possibles,
@@ -221,8 +239,9 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_diagonal_down_left(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
         possibles = queen.possible_positions_diagonal_down_left(4, 4)
         self.assertEqual(
             possibles,
@@ -230,9 +249,10 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_diagonal_down_left_with_obstacle(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
-        self.board.set_piece(2, 2, Pawn("WHITE", self.board))
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
+        board.set_piece(2, 2, Pawn("WHITE", board))
         possibles = queen.possible_positions_diagonal_down_left(4, 4)
         self.assertEqual(
             possibles,
@@ -240,9 +260,10 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_move_diagonal_down_left_attack(self):
-        queen = Queen("WHITE", self.board)
-        self.board.set_piece(4, 4, queen)
-        self.board.set_piece(2, 2, Pawn("BLACK", self.board))
+        board = Board(for_test=True)
+        queen = Queen("WHITE", board)
+        board.set_piece(4, 4, queen)
+        board.set_piece(2, 2, Pawn("BLACK", board))
         possibles = queen.possible_positions_diagonal_down_left(4, 4)
         self.assertEqual(
             possibles,
@@ -250,8 +271,9 @@ class TestQueen(unittest.TestCase):
         )
 
     def test_invalid_move(self):
-        self.board.set_piece(4, 4, Queen("WHITE", self.board))  
-        queen = self.board.get_piece(4, 4)  
+        board = Board(for_test=True)
+        board.set_piece(4, 4, Queen("WHITE", board))  
+        queen = board.get_piece(4, 4)  
         is_possible = queen.valid_positions(
             from_row=4,
             from_col=4,