+++ tests/test_rook.py
@@ -5,23 +5,29 @@ from game.pieces.pawn import Pawn
 
 class TestRook(unittest.TestCase):
 
-    def setUp(self):
-        self.board = Board()  # Crear un tablero con las piezas iniciales
-        # Limpiar el tablero
-        for row in range(8):
-            for col in range(8):
-                self.board.set_piece(row, col, None)
+    # def setUp(self):
+    #     board = Board()  # Crear un tablero con las piezas iniciales
+    #     # Limpiar el tablero
+    #     for row in range(8):
+    #         for col in range(8):
+    #             board.set_piece(row, col, None)
+
+    ####### No es necesario hacer esto, ya que con el 
+    # for_test=True
+    # logramos hacer lo mismo que buscamos #######
 
     def test_str(self):
-        rook = Rook("WHITE", self.board)
+        board = Board()
+        rook = Rook("WHITE", board)
         self.assertEqual(
             str(rook),
             "♜",
         )
 
     def test_move_vertically_down(self):
-        rook = Rook("WHITE", self.board)
-        self.board.set_piece(4, 1, rook)
+        board = Board(for_test=True)
+        rook = Rook("WHITE", board)
+        board.set_piece(4, 1, rook)
         possibles = rook.possible_movement_vertical_down(4, 1)
         self.assertEqual(
             possibles,
@@ -29,9 +35,10 @@ class TestRook(unittest.TestCase):
         )
 
     def test_move_vertically_down_with_obstacle(self):
-        rook = Rook("WHITE", self.board)
-        self.board.set_piece(4, 1, rook)
-        self.board.set_piece(2, 1, Pawn("WHITE", self.board))
+        board = Board(for_test=True)
+        rook = Rook("WHITE", board)
+        board.set_piece(4, 1, rook)
+        board.set_piece(2, 1, Pawn("WHITE", board))
         possibles = rook.possible_movement_vertical_down(4, 1)
         self.assertEqual(
             possibles,
@@ -39,9 +46,10 @@ class TestRook(unittest.TestCase):
         )
 
     def test_move_vertically_down_attack(self):
-        rook = Rook("WHITE", self.board)
-        self.board.set_piece(4, 1, rook)
-        self.board.set_piece(2, 1, Pawn("BLACK", self.board))
+        board = Board(for_test=True)
+        rook = Rook("WHITE", board)
+        board.set_piece(4, 1, rook)
+        board.set_piece(2, 1, Pawn("BLACK", board))
         possibles = rook.possible_movement_vertical_down(4, 1)
         self.assertEqual(
             possibles,
@@ -49,8 +57,9 @@ class TestRook(unittest.TestCase):
         )
 
     def test_move_vertically_up(self):
-        rook = Rook("WHITE", self.board)
-        self.board.set_piece(4, 1, rook)
+        board = Board(for_test=True)
+        rook = Rook("WHITE", board)
+        board.set_piece(4, 1, rook)
         possibles = rook.possible_movement_vertical_up(4, 1)
         self.assertEqual(
             possibles,
@@ -58,9 +67,10 @@ class TestRook(unittest.TestCase):
         )
 
     def test_move_vertically_up_with_obstacle(self):
-        rook = Rook("WHITE", self.board)
-        self.board.set_piece(4, 1, rook)
-        self.board.set_piece(6, 1, Pawn("WHITE", self.board))
+        board = Board(for_test=True)
+        rook = Rook("WHITE", board)
+        board.set_piece(4, 1, rook)
+        board.set_piece(6, 1, Pawn("WHITE", board))
         possibles = rook.possible_movement_vertical_up(4, 1)
         self.assertEqual(
             possibles,
@@ -68,9 +78,10 @@ class TestRook(unittest.TestCase):
         )
 
     def test_move_vertically_up_attack(self):
-        rook = Rook("WHITE", self.board)
-        self.board.set_piece(4, 1, rook)
-        self.board.set_piece(6, 1, Pawn("BLACK", self.board))
+        board = Board(for_test=True)
+        rook = Rook("WHITE", board)
+        board.set_piece(4, 1, rook)
+        board.set_piece(6, 1, Pawn("BLACK", board))
         possibles = rook.possible_movement_vertical_up(4, 1)
         self.assertEqual(
             possibles,
@@ -78,8 +89,9 @@ class TestRook(unittest.TestCase):
         )
 
     def test_move_horizontally_right(self):
-        rook = Rook("WHITE", self.board)
-        self.board.set_piece(4, 1, rook)
+        board = Board(for_test=True)
+        rook = Rook("WHITE", board)
+        board.set_piece(4, 1, rook)
         possibles = rook.possible_movement_horizontal_right(4, 1)
         self.assertEqual(
             possibles,
@@ -87,9 +99,10 @@ class TestRook(unittest.TestCase):
         )
 
     def test_move_horizontally_right_with_obstacle(self):
-        rook = Rook("WHITE", self.board)
-        self.board.set_piece(4, 1, rook)
-        self.board.set_piece(4, 3, Pawn("WHITE", self.board))
+        board = Board(for_test=True)
+        rook = Rook("WHITE", board)
+        board.set_piece(4, 1, rook)
+        board.set_piece(4, 3, Pawn("WHITE", board))
         possibles = rook.possible_movement_horizontal_right(4, 1)
         self.assertEqual(
             possibles,
@@ -97,9 +110,10 @@ class TestRook(unittest.TestCase):
         )
 
     def test_move_horizontally_right_attack(self):
-        rook = Rook("WHITE", self.board)
-        self.board.set_piece(4, 1, rook)
-        self.board.set_piece(4, 3, Pawn("BLACK", self.board))
+        board = Board(for_test=True)
+        rook = Rook("WHITE", board)
+        board.set_piece(4, 1, rook)
+        board.set_piece(4, 3, Pawn("BLACK", board))
         possibles = rook.possible_movement_horizontal_right(4, 1)
         self.assertEqual(
             possibles,
@@ -107,8 +121,9 @@ class TestRook(unittest.TestCase):
         )
 
     def test_move_horizontally_left(self):
-        rook = Rook("WHITE", self.board)
-        self.board.set_piece(4, 6, rook)
+        board = Board(for_test=True)
+        rook = Rook("WHITE", board)
+        board.set_piece(4, 6, rook)
         possibles = rook.possible_movement_horizontal_left(4, 6)
         self.assertEqual(
             possibles,
@@ -116,9 +131,10 @@ class TestRook(unittest.TestCase):
         )
 
     def test_move_horizontally_left_with_obstacle(self):
-        rook = Rook("WHITE", self.board)
-        self.board.set_piece(4, 6, rook)
-        self.board.set_piece(4, 4, Pawn("WHITE", self.board))
+        board = Board(for_test=True)
+        rook = Rook("WHITE", board)
+        board.set_piece(4, 6, rook)
+        board.set_piece(4, 4, Pawn("WHITE", board))
         possibles = rook.possible_movement_horizontal_left(4, 6)
         self.assertEqual(
             possibles,
@@ -126,9 +142,10 @@ class TestRook(unittest.TestCase):
         )
 
     def test_move_horizontally_left_attack(self):
-        rook = Rook("WHITE", self.board)
-        self.board.set_piece(4, 6, rook)
-        self.board.set_piece(4, 4, Pawn("BLACK", self.board))
+        board = Board(for_test=True)
+        rook = Rook("WHITE", board)
+        board.set_piece(4, 6, rook)
+        board.set_piece(4, 4, Pawn("BLACK", board))
         possibles = rook.possible_movement_horizontal_left(4, 6)
         self.assertEqual(
             possibles,
@@ -136,8 +153,9 @@ class TestRook(unittest.TestCase):
         )
 
     def test_invalid_move_diagonally(self):
-        self.board.set_piece(0, 0, Rook("WHITE", self.board))  # Colocar una torre en la posición (0, 0)
-        rook = self.board.get_piece(0, 0)  # Obtener la torre en la posición inicial
+        board = Board(for_test=True)
+        board.set_piece(0, 0, Rook("WHITE", board))  # Colocar una torre en la posición (0, 0)
+        rook = board.get_piece(0, 0)  # Obtener la torre en la posición inicial
         is_possible = rook.valid_positions(
             from_row=0,
             from_col=0,